+++ src/tpm2-command.c
@@ -270,7 +270,7 @@ tpm2_command_get_handle_count (Tpm2Command *command)
         return 0;
     }
     tmp = tpm2_command_get_attributes (command).val;
-    return (tmp & TPMA_CC_CHANDLES) >> 25;
+    return (guint8)((tmp & TPMA_CC_CHANDLES) >> 25);
 }
 /*
  * Simple function to access handles in the provided Tpm2Command. The
